pyproject.toml
[project]
name = "mybuddy"
version = "0.1.0"
description = "Personal assistant that collects notes, extracts action items and contacts with AI"
readme = "README.md"
authors = [
    { name = "kirancherlakola", email = "kiran.cherlakola@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mybuddy = "mybuddy.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
